backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
langchain
google-generativeai
python-dotenv
docx2txt
PyMuPDF
